hlsl: factor storage logic into a separate module
